+++ test/Sentry.Tests/Internals/SentryStackTraceFactoryTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Diagnostics;
 using System.Linq;
+using FluentAssertions;
 using Sentry;
 using Sentry.Extensibility;
 using Sentry.Protocol;
@@ -87,6 +88,24 @@ namespace Other.Tests.Internals
         }
 
         [Fact]
+        public void Create_ByRefFields_IncludesAmpersandAfterParameterType()
+        {
+            // Arrange
+            var i = 5;
+            var exception = Record.Exception(() => ByRefMethodThatThrows(i, in i, ref i, out i));
+
+            _fixture.SentryOptions.AttachStacktrace = true;
+            var factory = _fixture.GetSut();
+
+            // Act
+            var stackTrace = factory.Create(exception);
+
+            // Assert
+            var frame = stackTrace!.Frames.Last();
+            frame.Function.Should().Be("ByRefMethodThatThrows(Int32 value, Int32& valueIn, Int32& valueRef, Int32& valueOut)");
+        }
+
+        [Fact]
         public void CreateSentryStackFrame_AppNamespace_InAppFrame()
         {
             var frame = new StackFrame();
@@ -146,5 +165,8 @@ namespace Other.Tests.Internals
             SentryStackTraceFactory.DemangleAnonymousFunction(stackFrame);
             Assert.Null(stackFrame.Module);
         }
+
+        private static void ByRefMethodThatThrows(int value, in int valueIn, ref int valueRef, out int valueOut) =>
+            throw new Exception();
     }
 }